test(api): vitest+supertest の回帰テスト追加 / refactor: Express app を切り出し (src/app.js), SQLite in-memory 対応
